requirements.txt
Flask==2.3.2
flask-bcrypt==1.0.1
flask-session==0.4.0
python-dotenv==1.0.0

